refactor(processing): split processor.rs into orchestrator + helper-family modules
processor.rs (2293 LOC) carried the streaming orchestrator plus five cohesive
helper families. Convert it into a processor/ directory:

  mod.rs          process_geometry* entry points + EntityJob spine + orchestrator
  site_local.rs   mesh-space consts + site-local rotation/translation helpers
  properties.rs   space/zone property collection + metadata population
  quick_metadata.rs quick spatial-tree node parsing
  color_layer.rs  element-colour + presentation-layer resolution
  opening_filter.rs apply_opening_filter / is_opaque_opening / has_glass_style

Pure code motion. Helpers called from mod.rs gain pub(super); the EntityJob
and QuickSpatialNodeEntry spine structs + their fields gain pub(super); the
three *_MESH_COORDINATE_SPACE consts move to site_local.rs as pub(super) with
identical string values. The public API (process_geometry*, OpeningFilterMode,
ProcessingResult, StreamingOptions, convert_mesh_to_site_local,
get_refs_from_list) keeps its exact visibility and crate::processor::* paths;
lib.rs is untouched. The wasm32/native cfg split stays wholly inside mod.rs.
cargo test -p ifc-lite-processing green (21 binaries); clippy -D warnings clean;
wasm32-unknown-unknown build clean.
