+++ rust/processing/src/processor/mod.rs
@@ -9,10 +9,10 @@
 use crate::types::mesh::MeshData;
 use crate::types::response::{
     CoordinateInfo, ModelMetadata, ProcessingStats, QuickMetadataBootstrap,
-    QuickMetadataEntitySummary, QuickMetadataSpatialNode,
+    QuickMetadataEntitySummary,
 };
 use ifc_lite_core::{
-    build_entity_index, AttributeValue, DecodedEntity, EntityDecoder,
+    build_entity_index, DecodedEntity, EntityDecoder,
     EntityIndex, EntityScanner, IfcType,
 };
 use ifc_lite_geometry::TessellationQuality;
@@ -21,6 +21,34 @@ use rayon::prelude::*;
 use rustc_hash::{FxHashMap, FxHashSet};
 use std::collections::{BTreeMap, HashMap, HashSet};
 use std::sync::Arc;
+
+mod color_layer;
+mod opening_filter;
+mod properties;
+mod quick_metadata;
+mod site_local;
+
+pub use site_local::convert_mesh_to_site_local;
+
+use color_layer::{
+    collect_presentation_layer_assignments, resolve_element_color_for_product_definition_shape,
+    resolve_presentation_layer_for_product_definition_shape,
+};
+use opening_filter::apply_opening_filter;
+use properties::{
+    assign_space_zone_properties, collect_property_set_definition,
+    collect_rel_defines_by_properties_link, extract_property_name_and_value, PropertySetDefinition,
+    RelDefinesByPropertiesLink,
+};
+use quick_metadata::{
+    build_quick_spatial_tree_node, extract_name_from_args, extract_storey_elevation_from_args,
+    is_quick_spatial_type_ci, parse_step_arguments, parse_step_ref, parse_step_ref_list,
+    QuickSpatialNodeEntry,
+};
+use site_local::{
+    translation_is_nonidentity, MODEL_RTC_MESH_COORDINATE_SPACE, RAW_IFC_MESH_COORDINATE_SPACE,
+    SITE_LOCAL_MESH_COORDINATE_SPACE,
+};
 
 /// Wall-clock timer for diagnostic `ProcessingStats`. On wasm32
 /// `std::time::Instant::now()` traps ("time not implemented on this platform"),
@@ -126,127 +154,22 @@ impl Default for StreamingOptions {
     }
 }
 
-const SITE_LOCAL_MESH_COORDINATE_SPACE: &str = "site_local";
-const MODEL_RTC_MESH_COORDINATE_SPACE: &str = "model_rtc";
-const RAW_IFC_MESH_COORDINATE_SPACE: &str = "raw_ifc";
-
-/// Epsilon (metres) below which a placement translation is treated as identity.
-/// Avoids overriding a detected RTC anchor when `IfcSite` sits at the origin
-/// while the geometry itself carries large world coordinates.
-const PLACEMENT_IDENTITY_EPSILON: f64 = 1e-9;
-
-#[inline]
-fn translation_is_nonidentity(t: (f64, f64, f64)) -> bool {
-    t.0.abs() > PLACEMENT_IDENTITY_EPSILON
-        || t.1.abs() > PLACEMENT_IDENTITY_EPSILON
-        || t.2.abs() > PLACEMENT_IDENTITY_EPSILON
-}
-
-/// Apply the inverse of the site placement's 3×3 rotation to in-place `f32`
-/// triplets (positions or normals). Translation is handled separately via the
-/// router's `rtc_offset`; this only rotates vertices into the site-local axis
-/// frame when that frame is non-identity.
-fn apply_inverse_rotation_in_place(values: &mut [f32], column_major_matrix: &[f64]) {
-    if values.len() < 3 || column_major_matrix.len() < 16 {
-        return;
-    }
-
-    let r00 = column_major_matrix[0];
-    let r10 = column_major_matrix[1];
-    let r20 = column_major_matrix[2];
-    let r01 = column_major_matrix[4];
-    let r11 = column_major_matrix[5];
-    let r21 = column_major_matrix[6];
-    let r02 = column_major_matrix[8];
-    let r12 = column_major_matrix[9];
-    let r22 = column_major_matrix[10];
-
-    let is_identity = (r00 - 1.0).abs() < PLACEMENT_IDENTITY_EPSILON
-        && r10.abs() < PLACEMENT_IDENTITY_EPSILON
-        && r20.abs() < PLACEMENT_IDENTITY_EPSILON
-        && r01.abs() < PLACEMENT_IDENTITY_EPSILON
-        && (r11 - 1.0).abs() < PLACEMENT_IDENTITY_EPSILON
-        && r21.abs() < PLACEMENT_IDENTITY_EPSILON
-        && r02.abs() < PLACEMENT_IDENTITY_EPSILON
-        && r12.abs() < PLACEMENT_IDENTITY_EPSILON
-        && (r22 - 1.0).abs() < PLACEMENT_IDENTITY_EPSILON;
-    if is_identity {
-        return;
-    }
-
-    for chunk in values.chunks_exact_mut(3) {
-        let x = chunk[0] as f64;
-        let y = chunk[1] as f64;
-        let z = chunk[2] as f64;
-        chunk[0] = (r00 * x + r10 * y + r20 * z) as f32;
-        chunk[1] = (r01 * x + r11 * y + r21 * z) as f32;
-        chunk[2] = (r02 * x + r12 * y + r22 * z) as f32;
-    }
-}
-
-/// Rotate a mesh into the site-local axis frame. Only runs for the
-/// `site_local` coordinate-space tier; translation alignment happens upstream
-/// via the router's RTC subtraction.
-///
-/// Exposed so the streaming server can apply the same rotation to meshes it
-/// produces outside this crate's parallel loop.
-pub fn convert_mesh_to_site_local(mesh: &mut MeshData, site_transform: Option<&Vec<f64>>) {
-    let Some(site_transform) = site_transform else {
-        return;
-    };
-
-    apply_inverse_rotation_in_place(&mut mesh.positions, site_transform);
-    apply_inverse_rotation_in_place(&mut mesh.normals, site_transform);
-    // Positions are stored RELATIVE to `mesh.origin`, so the world point is
-    // `origin + position`. The site-local inverse rotation acts on the world
-    // point, so the origin must be rotated by the SAME inverse rotation (in f64)
-    // — otherwise the element would be rotated about the wrong centre.
-    apply_inverse_rotation_point_f64(&mut mesh.origin, site_transform);
-}
-
-/// Inverse-rotate a single f64 point in place by `column_major_matrix` (the same
-/// Rᵀ used by `apply_inverse_rotation_in_place`). Used for the per-mesh origin.
-fn apply_inverse_rotation_point_f64(p: &mut [f64; 3], column_major_matrix: &[f64]) {
-    if column_major_matrix.len() < 16 || (p[0] == 0.0 && p[1] == 0.0 && p[2] == 0.0) {
-        return;
-    }
-    let (r00, r10, r20) = (
-        column_major_matrix[0],
-        column_major_matrix[1],
-        column_major_matrix[2],
-    );
-    let (r01, r11, r21) = (
-        column_major_matrix[4],
-        column_major_matrix[5],
-        column_major_matrix[6],
-    );
-    let (r02, r12, r22) = (
-        column_major_matrix[8],
-        column_major_matrix[9],
-        column_major_matrix[10],
-    );
-    let (x, y, z) = (p[0], p[1], p[2]);
-    p[0] = r00 * x + r10 * y + r20 * z;
-    p[1] = r01 * x + r11 * y + r21 * z;
-    p[2] = r02 * x + r12 * y + r22 * z;
-}
-
 /// Job for processing a single entity.
-struct EntityJob {
-    id: u32,
-    ifc_type: IfcType,
-    start: usize,
-    end: usize,
-    product_definition_shape_id: Option<u32>,
-    element_color: [f32; 4],
-    global_id: Option<String>,
-    name: Option<String>,
-    presentation_layer: Option<String>,
-    space_zone_properties: Option<BTreeMap<String, String>>,
+pub(super) struct EntityJob {
+    pub(super) id: u32,
+    pub(super) ifc_type: IfcType,
+    pub(super) start: usize,
+    pub(super) end: usize,
+    pub(super) product_definition_shape_id: Option<u32>,
+    pub(super) element_color: [f32; 4],
+    pub(super) global_id: Option<String>,
+    pub(super) name: Option<String>,
+    pub(super) presentation_layer: Option<String>,
+    pub(super) space_zone_properties: Option<BTreeMap<String, String>>,
     /// Set for synthetic type-only-geometry jobs (#957): the `IfcRepresentationMap`
     /// id to render directly (baking its MappingOrigin) instead of walking the
     /// element's `IfcProductDefinitionShape`. `None` for ordinary product jobs.
-    representation_map_id: Option<u32>,
+    pub(super) representation_map_id: Option<u32>,
 }
 
 // Only invoked on the wasm32 serial path; dead on the native build.
@@ -315,18 +238,6 @@ fn populate_entity_job_metadata(
 // (via `from_color`) the browser batch path.
 use crate::style::GeometryStyleInfo;
 
-#[derive(Debug, Clone)]
-struct PropertySetDefinition {
-    name: Option<String>,
-    property_ids: Vec<u32>,
-}
-
-#[derive(Debug, Clone)]
-struct RelDefinesByPropertiesLink {
-    property_set_id: u32,
-    related_object_ids: Vec<u32>,
-}
-
 /// Extract entity references from a list attribute.
 pub(crate) fn get_refs_from_list(entity: &DecodedEntity, index: usize) -> Option<Vec<u32>> {
     let list = entity.get_list(index)?;
@@ -338,411 +249,12 @@ pub(crate) fn get_refs_from_list(entity: &DecodedEntity, index: usize) -> Option
     }
 }
 
-fn normalize_optional_string(raw: Option<&str>) -> Option<String> {
+pub(super) fn normalize_optional_string(raw: Option<&str>) -> Option<String> {
     let value = raw?.trim();
     if value.is_empty() || value == "$" {
         return None;
     }
     Some(value.to_string())
-}
-
-fn normalize_ifc_property_name(raw: Option<&str>) -> Option<String> {
-    let name = normalize_optional_string(raw)?;
-    let cleaned = name.trim();
-    if cleaned.is_empty() {
-        return None;
-    }
-
-    Some(cleaned.to_string())
-}
-
-fn is_space_or_zone_type(ifc_type: &IfcType) -> bool {
-    matches!(
-        ifc_type,
-        IfcType::IfcSpace
-            | IfcType::IfcSpaceType
-            | IfcType::IfcZone
-            | IfcType::IfcSpatialZone
-            | IfcType::IfcSpatialZoneType
-    )
-}
-
-fn collect_property_set_definition(property_set: &DecodedEntity) -> Option<PropertySetDefinition> {
-    let property_ids = property_set
-        .get_list(4)
-        .or_else(|| property_set.get_list(2))
-        .map(|items| {
-            items
-                .iter()
-                .filter_map(AttributeValue::as_entity_ref)
-                .collect::<Vec<u32>>()
-        })
-        .unwrap_or_default();
-
-    if property_ids.is_empty() {
-        return None;
-    }
-
-    let name = normalize_optional_string(property_set.get_string(2))
-        .or_else(|| normalize_optional_string(property_set.get_string(0)));
-
-    Some(PropertySetDefinition { name, property_ids })
-}
-
-fn collect_rel_defines_by_properties_link(
-    rel_defines: &DecodedEntity,
-) -> Option<RelDefinesByPropertiesLink> {
-    let property_set_id = rel_defines.get_ref(5).or_else(|| rel_defines.get_ref(3))?;
-    let related_object_ids = rel_defines
-        .get_list(4)
-        .or_else(|| rel_defines.get_list(2))
-        .map(|items| {
-            items
-                .iter()
-                .filter_map(AttributeValue::as_entity_ref)
-                .collect::<Vec<u32>>()
-        })
-        .unwrap_or_default();
-
-    if related_object_ids.is_empty() {
-        return None;
-    }
-
-    Some(RelDefinesByPropertiesLink {
-        property_set_id,
-        related_object_ids,
-    })
-}
-
-fn attribute_list_to_string(values: &[AttributeValue]) -> Option<String> {
-    let tokens = values
-        .iter()
-        .filter_map(attribute_value_to_string)
-        .collect::<Vec<String>>();
-
-    if tokens.is_empty() {
-        return None;
-    }
-
-    Some(tokens.join("; "))
-}
-
-fn attribute_value_to_string(value: &AttributeValue) -> Option<String> {
-    match value {
-        AttributeValue::Null | AttributeValue::Derived => None,
-        AttributeValue::String(text) => normalize_optional_string(Some(text)),
-        AttributeValue::Enum(text) => normalize_optional_string(Some(text.trim_matches('.'))),
-        AttributeValue::Integer(number) => Some(number.to_string()),
-        AttributeValue::Float(number) => Some(number.to_string()),
-        AttributeValue::EntityRef(id) => Some(format!("#{id}")),
-        AttributeValue::List(values) => {
-            if values.len() >= 2 && matches!(values.first(), Some(AttributeValue::String(_))) {
-                return values.get(1).and_then(attribute_value_to_string);
-            }
-
-            attribute_list_to_string(values)
-        }
-    }
-}
-
-fn extract_property_name_and_value(property_entity: &DecodedEntity) -> Option<(String, String)> {
-    let property_name = normalize_ifc_property_name(property_entity.get_string(0))
-        .or_else(|| normalize_ifc_property_name(property_entity.get_string(2)))?;
-
-    let property_type = property_entity.ifc_type.name();
-    let value = match property_type {
-        "IfcPropertySingleValue" => property_entity.get(2).and_then(attribute_value_to_string),
-        "IfcPropertyEnumeratedValue" => property_entity.get(2).and_then(attribute_value_to_string),
-        "IfcPropertyListValue" => property_entity.get(2).and_then(attribute_value_to_string),
-        "IfcPropertyBoundedValue" => {
-            let lower = property_entity.get(2).and_then(attribute_value_to_string);
-            let upper = property_entity.get(3).and_then(attribute_value_to_string);
-            match (lower, upper) {
-                (Some(lo), Some(hi)) => Some(format!("{lo}..{hi}")),
-                (Some(lo), None) => Some(lo),
-                (None, Some(hi)) => Some(hi),
-                (None, None) => None,
-            }
-        }
-        "IfcPropertyReferenceValue" => property_entity.get(2).and_then(attribute_value_to_string),
-        _ => None,
-    }?;
-
-    let normalized_value = value.trim();
-    if normalized_value.is_empty() || normalized_value == "$" {
-        return None;
-    }
-
-    Some((property_name, normalized_value.to_string()))
-}
-
-fn add_space_zone_property(
-    attributes: &mut BTreeMap<String, String>,
-    property_set_name: Option<&str>,
-    property_name: &str,
-    property_value: &str,
-) {
-    if property_name.trim().is_empty() || property_value.trim().is_empty() {
-        return;
-    }
-
-    attributes
-        .entry(property_name.to_string())
-        .or_insert_with(|| property_value.to_string());
-
-    if let Some(pset_name) = normalize_optional_string(property_set_name) {
-        let scoped_name = format!("{}.{}", pset_name, property_name);
-        attributes
-            .entry(scoped_name)
-            .or_insert_with(|| property_value.to_string());
-    }
-}
-
-fn build_space_zone_properties_by_entity(
-    entity_jobs: &[EntityJob],
-    property_values_by_id: &FxHashMap<u32, (String, String)>,
-    property_sets_by_id: &FxHashMap<u32, PropertySetDefinition>,
-    rel_defines_by_properties: &[RelDefinesByPropertiesLink],
-) -> FxHashMap<u32, BTreeMap<String, String>> {
-    let mut target_space_zone_ids = FxHashMap::default();
-    for job in entity_jobs
-        .iter()
-        .filter(|job| is_space_or_zone_type(&job.ifc_type))
-    {
-        target_space_zone_ids.insert(job.id, ());
-    }
-
-    if target_space_zone_ids.is_empty() {
-        return FxHashMap::default();
-    }
-
-    let mut properties_by_entity: FxHashMap<u32, BTreeMap<String, String>> = FxHashMap::default();
-
-    for link in rel_defines_by_properties {
-        let Some(property_set) = property_sets_by_id.get(&link.property_set_id) else {
-            continue;
-        };
-
-        for related_id in &link.related_object_ids {
-            if !target_space_zone_ids.contains_key(related_id) {
-                continue;
-            }
-
-            let attributes = properties_by_entity.entry(*related_id).or_default();
-            for property_id in &property_set.property_ids {
-                let Some((property_name, property_value)) = property_values_by_id.get(property_id)
-                else {
-                    continue;
-                };
-
-                add_space_zone_property(
-                    attributes,
-                    property_set.name.as_deref(),
-                    property_name,
-                    property_value,
-                );
-            }
-        }
-    }
-
-    properties_by_entity
-}
-
-fn assign_space_zone_properties(
-    entity_jobs: &mut [EntityJob],
-    property_values_by_id: &FxHashMap<u32, (String, String)>,
-    property_sets_by_id: &FxHashMap<u32, PropertySetDefinition>,
-    rel_defines_by_properties: &[RelDefinesByPropertiesLink],
-) {
-    let properties_by_entity = build_space_zone_properties_by_entity(
-        entity_jobs,
-        property_values_by_id,
-        property_sets_by_id,
-        rel_defines_by_properties,
-    );
-
-    if properties_by_entity.is_empty() {
-        return;
-    }
-
-    for job in entity_jobs.iter_mut() {
-        if let Some(properties) = properties_by_entity.get(&job.id) {
-            job.space_zone_properties = Some(properties.clone());
-        }
-    }
-}
-
-#[derive(Clone)]
-struct QuickSpatialNodeEntry {
-    express_id: u32,
-    type_name: String,
-    name: String,
-    elevation: Option<f64>,
-    children: Vec<u32>,
-    elements: Vec<u32>,
-    parent: Option<u32>,
-}
-
-/// Case-insensitive spatial-type check that avoids to_ascii_uppercase() allocation.
-#[inline]
-fn is_quick_spatial_type_ci(type_name: &str) -> bool {
-    type_name.eq_ignore_ascii_case("IFCPROJECT")
-        || type_name.eq_ignore_ascii_case("IFCSITE")
-        || type_name.eq_ignore_ascii_case("IFCBUILDING")
-        || type_name.eq_ignore_ascii_case("IFCBUILDINGSTOREY")
-        || type_name.eq_ignore_ascii_case("IFCSPACE")
-        || type_name.eq_ignore_ascii_case("IFCSPATIALZONE")
-        || type_name.eq_ignore_ascii_case("IFCFACILITY")
-        || type_name.eq_ignore_ascii_case("IFCFACILITYPART")
-        || type_name.eq_ignore_ascii_case("IFCBRIDGE")
-        || type_name.eq_ignore_ascii_case("IFCBRIDGEPART")
-        || type_name.eq_ignore_ascii_case("IFCROAD")
-        || type_name.eq_ignore_ascii_case("IFCROADPART")
-        || type_name.eq_ignore_ascii_case("IFCRAILWAY")
-        || type_name.eq_ignore_ascii_case("IFCRAILWAYPART")
-}
-
-fn parse_step_arguments(entity_bytes: &[u8]) -> Vec<&[u8]> {
-    let Some(open_idx) = entity_bytes.iter().position(|byte| *byte == b'(') else {
-        return Vec::new();
-    };
-    let Some(close_idx) = entity_bytes.iter().rposition(|byte| *byte == b')') else {
-        return Vec::new();
-    };
-    if close_idx <= open_idx {
-        return Vec::new();
-    }
-    let args = &entity_bytes[open_idx + 1..close_idx];
-    let mut parts = Vec::new();
-    let mut in_string = false;
-    let mut depth = 0i32;
-    let mut start = 0usize;
-    let bytes = args;
-    let mut index = 0usize;
-    while index < bytes.len() {
-        match bytes[index] {
-            b'\'' => {
-                if in_string && index + 1 < bytes.len() && bytes[index + 1] == b'\'' {
-                    index += 1;
-                } else {
-                    in_string = !in_string;
-                }
-            }
-            b'(' if !in_string => depth += 1,
-            b')' if !in_string => depth -= 1,
-            b',' if !in_string && depth == 0 => {
-                parts.push(args[start..index].trim_ascii());
-                start = index + 1;
-            }
-            _ => {}
-        }
-        index += 1;
-    }
-    if start <= args.len() {
-        parts.push(args[start..].trim_ascii());
-    }
-    parts
-}
-
-fn parse_step_string(token: &[u8]) -> Option<String> {
-    let trimmed = token.trim_ascii();
-    if trimmed.len() < 2 || trimmed[0] != b'\'' || trimmed[trimmed.len() - 1] != b'\'' {
-        return None;
-    }
-    let unescaped = String::from_utf8_lossy(&trimmed[1..trimmed.len() - 1]).replace("''", "'");
-    // Decode STEP unicode escapes so quick-metadata names match the from_token
-    // path and the TS parser (e.g. a name stored as Br\X2\00FC\X0\cke).
-    Some(ifc_lite_core::decode_ifc_string(&unescaped).into_owned())
-}
-
-fn parse_step_ref(token: &[u8]) -> Option<u32> {
-    std::str::from_utf8(token.trim_ascii().strip_prefix(b"#")?)
-        .ok()?
-        .parse()
-        .ok()
-}
-
-fn parse_step_ref_list(token: &[u8]) -> Vec<u32> {
-    let trimmed = token.trim_ascii();
-    let inner = trimmed
-        .strip_prefix(b"(")
-        .and_then(|value| value.strip_suffix(b")"))
-        .unwrap_or(trimmed);
-    inner.split(|byte| *byte == b',').filter_map(parse_step_ref).collect()
-}
-
-fn extract_name_from_args(args: &[&[u8]], fallback: &str) -> String {
-    args.get(2)
-        .and_then(|token| parse_step_string(token))
-        .filter(|value| !value.trim().is_empty())
-        .unwrap_or_else(|| fallback.to_string())
-}
-
-fn extract_storey_elevation_from_args(args: &[&[u8]]) -> Option<f64> {
-    for index in [9usize, 8usize] {
-        if let Some(value) = args
-            .get(index)
-            .and_then(|token| std::str::from_utf8(token.trim_ascii()).ok())
-            .and_then(|token| token.parse::<f64>().ok())
-        {
-            return Some(value);
-        }
-    }
-    args.iter()
-        .filter_map(|token| std::str::from_utf8(token.trim_ascii()).ok())
-        .filter_map(|token| token.parse::<f64>().ok())
-        .find(|value| value.abs() < 10_000.0)
-}
-
-fn build_quick_spatial_tree_node(
-    express_id: u32,
-    nodes: &HashMap<u32, QuickSpatialNodeEntry>,
-    element_summaries: &HashMap<u32, QuickMetadataEntitySummary>,
-) -> Result<QuickMetadataSpatialNode, String> {
-    let node = nodes
-        .get(&express_id)
-        .ok_or_else(|| format!("Quick spatial node #{express_id} not found"))?;
-    let mut children = Vec::with_capacity(node.children.len());
-    for child_id in &node.children {
-        children.push(build_quick_spatial_tree_node(
-            *child_id,
-            nodes,
-            element_summaries,
-        )?);
-    }
-    let elements = node
-        .elements
-        .iter()
-        .map(|element_id| {
-            element_summaries
-                .get(element_id)
-                .cloned()
-                .unwrap_or(QuickMetadataEntitySummary {
-                express_id: *element_id,
-                type_name: "IfcProduct".to_string(),
-                name: format!("IfcProduct #{}", element_id),
-                global_id: None,
-                kind: "element".to_string(),
-                has_children: false,
-                element_count: None,
-                elevation: None,
-            })
-        })
-        .collect();
-    Ok(QuickMetadataSpatialNode {
-        summary: QuickMetadataEntitySummary {
-            express_id: node.express_id,
-            type_name: node.type_name.clone(),
-            name: node.name.clone(),
-            global_id: None,
-            kind: "spatial".to_string(),
-            has_children: !node.children.is_empty() || !node.elements.is_empty(),
-            element_count: Some(node.elements.len()),
-            elevation: node.elevation,
-        },
-        children,
-        elements,
-    })
 }
 
 fn geometry_priority_score(ifc_type: &IfcType) -> u8 {
@@ -1899,379 +1411,6 @@ fn build_color_updates_for_jobs(
     }
 
     updates
-}
-
-fn collect_presentation_layer_assignments(
-    layer_by_assigned_representation: &mut FxHashMap<u32, String>,
-    layer_assignment: &DecodedEntity,
-) {
-    let Some(layer_name) = normalize_optional_string(layer_assignment.get_string(0)) else {
-        return;
-    };
-
-    let Some(assigned_items) = get_refs_from_list(layer_assignment, 2) else {
-        return;
-    };
-
-    for assigned in assigned_items {
-        layer_by_assigned_representation
-            .entry(assigned)
-            .or_insert_with(|| layer_name.clone());
-    }
-}
-
-fn resolve_element_color_for_product_definition_shape(
-    product_definition_shape_id: u32,
-    geometry_styles: &FxHashMap<u32, GeometryStyleInfo>,
-    decoder: &mut EntityDecoder,
-) -> Option<[f32; 4]> {
-    find_color_in_representation(product_definition_shape_id, geometry_styles, decoder)
-}
-
-fn resolve_presentation_layer_for_product_definition_shape(
-    product_definition_shape_id: u32,
-    layer_by_assigned_representation: &FxHashMap<u32, String>,
-    cache_by_representation: &mut FxHashMap<u32, Option<String>>,
-    decoder: &mut EntityDecoder,
-) -> Option<String> {
-    if let Some(layer_name) = layer_by_assigned_representation.get(&product_definition_shape_id) {
-        return Some(layer_name.clone());
-    }
-
-    let product_definition_shape = decoder.decode_by_id(product_definition_shape_id).ok()?;
-    let representation_ids = get_refs_from_list(&product_definition_shape, 2)?;
-
-    for representation_id in representation_ids {
-        if let Some(layer_name) = resolve_presentation_layer_name(
-            representation_id,
-            layer_by_assigned_representation,
-            cache_by_representation,
-            decoder,
-            &mut Vec::new(),
-        ) {
-            return Some(layer_name);
-        }
-    }
-
-    None
-}
-
-fn resolve_presentation_layer_name(
-    representation_id: u32,
-    layer_by_assigned_representation: &FxHashMap<u32, String>,
-    cache_by_representation: &mut FxHashMap<u32, Option<String>>,
-    decoder: &mut EntityDecoder,
-    traversal_stack: &mut Vec<u32>,
-) -> Option<String> {
-    if let Some(cached) = cache_by_representation.get(&representation_id) {
-        return cached.clone();
-    }
-
-    if traversal_stack.contains(&representation_id) {
-        return None;
-    }
-    traversal_stack.push(representation_id);
-
-    if let Some(layer_name) = layer_by_assigned_representation.get(&representation_id) {
-        let result = Some(layer_name.clone());
-        cache_by_representation.insert(representation_id, result.clone());
-        traversal_stack.pop();
-        return result;
-    }
-
-    let mut resolved: Option<String> = None;
-
-    if let Ok(representation) = decoder.decode_by_id(representation_id) {
-        if let Some(items) = get_refs_from_list(&representation, 3) {
-            for item_id in items {
-                if let Some(layer_name) = layer_by_assigned_representation.get(&item_id) {
-                    resolved = Some(layer_name.clone());
-                    break;
-                }
-
-                if let Ok(item) = decoder.decode_by_id(item_id) {
-                    if item.ifc_type == IfcType::IfcMappedItem {
-                        if let Some(mapping_source_id) = item.get_ref(0) {
-                            if let Ok(mapping_source) = decoder.decode_by_id(mapping_source_id) {
-                                if let Some(mapped_representation_id) = mapping_source.get_ref(1) {
-                                    if let Some(layer_name) = resolve_presentation_layer_name(
-                                        mapped_representation_id,
-                                        layer_by_assigned_representation,
-                                        cache_by_representation,
-                                        decoder,
-                                        traversal_stack,
-                                    ) {
-                                        resolved = Some(layer_name);
-                                        break;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    traversal_stack.pop();
-    cache_by_representation.insert(representation_id, resolved.clone());
-    resolved
-}
-
-/// Find a color in a representation by traversing its items.
-fn find_color_in_representation(
-    repr_id: u32,
-    geometry_styles: &FxHashMap<u32, GeometryStyleInfo>,
-    decoder: &mut EntityDecoder,
-) -> Option<[f32; 4]> {
-    // Decode the IfcProductDefinitionShape
-    let repr = decoder.decode_by_id(repr_id).ok()?;
-
-    // Attribute 2: Representations (list of IfcRepresentation)
-    let repr_list = get_refs_from_list(&repr, 2)?;
-
-    for shape_repr_id in repr_list {
-        if let Ok(shape_repr) = decoder.decode_by_id(shape_repr_id) {
-            // Attribute 3: Items (list of IfcRepresentationItem)
-            if let Some(items) = get_refs_from_list(&shape_repr, 3) {
-                for item_id in items {
-                    // Check direct style
-                    if let Some(style) = geometry_styles.get(&item_id) {
-                        return Some(style.color);
-                    }
-
-                    // Check mapped items
-                    if let Ok(item) = decoder.decode_by_id(item_id) {
-                        if item.ifc_type == IfcType::IfcMappedItem {
-                            if let Some(source_id) = item.get_ref(0) {
-                                if let Ok(source) = decoder.decode_by_id(source_id) {
-                                    if let Some(mapped_repr_id) = source.get_ref(1) {
-                                        if let Some(color) = find_color_in_shape_representation(
-                                            mapped_repr_id,
-                                            geometry_styles,
-                                            decoder,
-                                        ) {
-                                            return Some(color);
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    None
-}
-
-/// Find color in a shape representation.
-fn find_color_in_shape_representation(
-    repr_id: u32,
-    geometry_styles: &FxHashMap<u32, GeometryStyleInfo>,
-    decoder: &mut EntityDecoder,
-) -> Option<[f32; 4]> {
-    let repr = decoder.decode_by_id(repr_id).ok()?;
-    let items = get_refs_from_list(&repr, 3)?;
-
-    for item_id in items {
-        if let Some(style) = geometry_styles.get(&item_id) {
-            return Some(style.color);
-        }
-    }
-
-    None
-}
-
-/// Apply the opening filter and return which entity IDs to suppress and a filtered void index.
-///
-/// Returns `(skipped_entity_ids, filtered_void_index)` where:
-/// - `skipped_entity_ids` is the set of IfcWindow/IfcDoor entity IDs to omit from geometry output
-/// - `filtered_void_index` is the void index with suppressed openings removed from host lists
-fn apply_opening_filter(
-    entity_jobs: &[EntityJob],
-    void_index: &FxHashMap<u32, Vec<u32>>,
-    filling_by_opening: &FxHashMap<u32, u32>,
-    geometry_style_index: &FxHashMap<u32, GeometryStyleInfo>,
-    decoder: &mut EntityDecoder,
-    mode: OpeningFilterMode,
-) -> (HashSet<u32>, FxHashMap<u32, Vec<u32>>) {
-    if mode == OpeningFilterMode::Default {
-        return (HashSet::default(), void_index.clone());
-    }
-
-    // Collect all IfcWindow / IfcDoor entity jobs.
-    let filling_jobs: FxHashMap<u32, &EntityJob> = entity_jobs
-        .iter()
-        .filter(|job| matches!(job.ifc_type, IfcType::IfcWindow | IfcType::IfcDoor))
-        .map(|job| (job.id, job))
-        .collect();
-
-    if filling_jobs.is_empty() {
-        return (HashSet::default(), void_index.clone());
-    }
-
-    let mut skipped_entity_ids: HashSet<u32> = HashSet::default();
-
-    // IgnoreAll: suppress every window/door mesh and clear ALL wall voids.
-    // We always clear the full void_index because IfcRelFillsElement is often absent
-    // or only partially present, and without it we cannot identify which specific openings
-    // belong to windows/doors.
-    if mode == OpeningFilterMode::IgnoreAll {
-        for &id in filling_jobs.keys() {
-            skipped_entity_ids.insert(id);
-        }
-        return (skipped_entity_ids, FxHashMap::default());
-    }
-
-    // IgnoreOpaque: suppress only windows/doors that have no transparent sub-parts.
-    // Mesh suppression uses element color + style traversal (is_opaque_opening).
-    // Void suppression uses IfcRelFillsElement data when available.
-    for (&id, job) in &filling_jobs {
-        if is_opaque_opening(job, geometry_style_index, decoder) {
-            skipped_entity_ids.insert(id);
-        }
-    }
-
-    if filling_by_opening.is_empty() {
-        // No IfcRelFillsElement — can't map voids to specific window/door entities.
-        return (skipped_entity_ids, void_index.clone());
-    }
-
-    // Build openings_to_suppress from the explicit opening → filling mapping.
-    let mut openings_to_suppress: HashSet<u32> = HashSet::default();
-    for (&opening_id, &filling_id) in filling_by_opening {
-        if skipped_entity_ids.contains(&filling_id) {
-            openings_to_suppress.insert(opening_id);
-        }
-    }
-
-    if openings_to_suppress.is_empty() {
-        return (skipped_entity_ids, void_index.clone());
-    }
-
-    let mut filtered: FxHashMap<u32, Vec<u32>> = FxHashMap::default();
-    for (&host_id, openings) in void_index {
-        let remaining: Vec<u32> = openings
-            .iter()
-            .copied()
-            .filter(|oid| !openings_to_suppress.contains(oid))
-            .collect();
-        if !remaining.is_empty() {
-            filtered.insert(host_id, remaining);
-        }
-    }
-
-    (skipped_entity_ids, filtered)
-}
-
-/// Returns `true` when the entity has no transparent or glass sub-parts,
-/// meaning it is an opaque window/door that should be suppressed by `IgnoreOpaque`.
-///
-/// Any of the following makes it NOT opaque (returns `false`):
-/// - Entity name contains "glas" (case-insensitive)
-/// - Resolved element color has any transparency (alpha < 1.0)
-/// - Any sub-geometry style has alpha < 1.0 or a material/style name containing "glas"
-fn is_opaque_opening(
-    job: &EntityJob,
-    styles: &FxHashMap<u32, GeometryStyleInfo>,
-    decoder: &mut EntityDecoder,
-) -> bool {
-    let Ok(entity) = decoder.decode_at(job.start, job.end) else {
-        return true;
-    };
-
-    // 1. Entity name contains "glas" → glazed.
-    if normalize_optional_string(entity.get_string(2))
-        .as_deref()
-        .map(|n| n.to_lowercase().contains("glas"))
-        .unwrap_or(false)
-    {
-        return false;
-    }
-
-    // 2. Resolved element color has any transparency → glazed.
-    //    Covers IfcWindow entities using their default colour ([0.6, 0.8, 1.0, 0.4])
-    //    and any entity whose explicit surface style resolved to a transparent colour.
-    if job.element_color[3] < 1.0 {
-        return false;
-    }
-
-    let Some(product_shape_id) = entity.get_ref(6) else {
-        return true; // No shape info — treat as opaque
-    };
-
-    let Ok(product_shape) = decoder.decode_by_id(product_shape_id) else {
-        return true;
-    };
-
-    let Some(repr_ids) = get_refs_from_list(&product_shape, 2) else {
-        return true;
-    };
-
-    for repr_id in repr_ids {
-        let Ok(repr) = decoder.decode_by_id(repr_id) else {
-            continue;
-        };
-        let Some(item_ids) = get_refs_from_list(&repr, 3) else {
-            continue;
-        };
-        for item_id in item_ids {
-            // Direct style on item
-            if let Some(style) = styles.get(&item_id) {
-                if has_glass_style(style) {
-                    return false;
-                }
-            }
-
-            // Mapped items: IfcMappedItem → IfcRepresentationMap → IfcRepresentation → items
-            if let Ok(item) = decoder.decode_by_id(item_id) {
-                if item.ifc_type == IfcType::IfcMappedItem {
-                    if let Some(source_id) = item.get_ref(0) {
-                        if let Ok(source) = decoder.decode_by_id(source_id) {
-                            if let Some(mapped_repr_id) = source.get_ref(1) {
-                                if let Ok(mapped_repr) = decoder.decode_by_id(mapped_repr_id) {
-                                    if let Some(mapped_items) = get_refs_from_list(&mapped_repr, 3)
-                                    {
-                                        for mapped_item_id in mapped_items {
-                                            if let Some(style) = styles.get(&mapped_item_id) {
-                                                if has_glass_style(style) {
-                                                    return false;
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    true // No glass found → opaque
-}
-
-/// Returns `true` when a geometry style indicates a glass/transparent material.
-///
-/// Triggers on:
-/// - Any transparency at all (alpha < 1.0)
-/// - Style/material name containing "glas" (case-insensitive)
-fn has_glass_style(style: &GeometryStyleInfo) -> bool {
-    if style.color[3] < 1.0 {
-        return true;
-    }
-    if style
-        .material_name
-        .as_deref()
-        .map(|n| n.to_lowercase().contains("glas"))
-        .unwrap_or(false)
-    {
-        return true;
-    }
-    false
 }
